[move] inscriptions for Aptos
This is a quick demo of one way in which we could do inscriptions on
Aptos.

The details:
* Builds upon our digital asset standard (token objects)
* inscription obtains a unique global id beginning from 0
* inscriptions store data as events
* current inscriptions are limited to the transaction size limit of 64K,
  so like 62K images

This includes an immutable collection inscription contract for some
initial minting. The expectation is that launchpads will build their own
way to create inscriptions, but this immutable collection contract works
pretty well for phase 1 inscriptions.

I would love feedback about how we can make this standard better. Should
we include content type? We definitely need a demo using the on-chain
indexers for testnet -- but let's first agree on the contract itself.

I'll push out an AIP on this topic very shortly for further
conversations, but for now, let's start on this PR.